rename scripts; block from running outside CI

+++ scripts/compare_ci_diff_types.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! "$CI" = "true" ] || [ ! -f /.dockerenv ]; then
+    echo `basename $0` "can run only in Circle CI"
+    exit 1
+fi
+
 # cleanup if needed
 
 git clean -dfx
